Preprocessing is handled after lexing now, flush out basic preprocessor.
The preprocessor always has to tokenize it's input so it's much easier
to just tokenize everything beforehand.  This means that a Token can
be a "Macro" (which is just a Vec<Tag<Token>>).

The lexer now uses a "prelexer" to handle deleting escaped line breaks
and comments.

The combination of these changes makes it so easy to write the code
and avoids work duplication.